Refactor tabs
